Add lint script for CI
Based on the system used in https://github.com/zivid/zivid-python, but
since we only use it for linting for now, we only support Ubuntu 18.04.
When/if we start running more involved tests, we'll add support for
other OSes as well.

This is not actually run in CI yet, can be run locally with this
command:

docker run --volume $PWD:/host --workdir /host/continuous-integration/linux ubuntu:18.04 bash -c "./setup.sh && ./lint.sh"