Don't look for packages if we're not using them
Moving `find_package` inside `if(USE...)` makes sure we don't call them
if we're not using them.

I'm still keeping them as optional though (not adding `REQUIRED`), so
that we can also print our own error message informing the user that
they have the option to turn the dependency off.