Enable warnings
This commit sets up the infrastructure to enable more or less all
warnings on Clang, GCC and MSVC. It also disables compiler extensions,
to ensure portability.

Warnings are disabled in libraries, so we don't get warnings
from for instance Eigen and PCL.

In CI, the plan is for warnings to be enabled and treated as errors.
Currently we do not treat them as errors, but when all warnings in the
current code base have been fixed, we should set -DWARNINGS_AS_ERRORS=ON
in build.sh.

By default, warnings are disabled, to make it easier for users to modify
and play with the code. Contributors to this repo are encouraged to
enable WARNINGS and WARNINGS_AS_ERRORS when configuring CMake, so they
don't have to wait for CI to tell them.

Some warnings we never want to turn on. If a particular warning should
be disabled for a particular compiler, add it to
WARNINGS_THAT_SHOULD_BE_IGNORED for that compiler in
CompilerOptions.cmake. You must make a note of why it should be ignored.
I've added a few of them for Clang, to demonstrate how it's done.
Someone should do the same for GCC and MSVC.

I've not gone through and fixed all warnings in the code base yet.

Issue #36