CI: Use correct source and build dirs
build.sh used to build in the directory `/source/build` relative to the
repo root. Now we instead build in `/build` relative to the repo root.
This was the intention all along, the previous was just a bug.

lint.sh used to look for files in the entire repo. It now only looks in
the source folder, in order to avoid finding generated files in the
build folder.

This fixes #65