Disable hicpp-signed-bitwise warnings in clang-tidy
We get a lot of these warnings from OpenCV. So let's ignore it in
.clang-tidy instead of littering the samples with explicit ignores.

Details:

The issue is that OpenCV decided to use a signed integer for this, which
they also do bitwise operations on. That can often lead to undefined or
implementation defined behaviour. But it's outside of our control to fix
this.