3dTo2d projection with OpenCV
Simply run the sample with a connected camera. It will run through
a number of different capture settings. For each setting it will
project the pointcloud back into the camera, and calculate statistics
based on the pixel offsets. There is also a heatmap that visualizes it.
Disable the visualization (the images are still saved) by commenting out
lines with cv::imshow() and cv::waitKey(0)

I chose COLORMAP_HSV because it has a narrow range around the center.
COLORMAP_PARULA would also be nice as it has a green center range and
clear distinction between positive and negative pixel offset.

The pixel offsets are scaled to UINT8, with center around 128. To be
able compare between settings I used a fixed range of +-20 pixels. Note
that offsets above this is clipped. What I found is that it is hard to
trigger such large offsets, so it's a bit difficult to see.