Samples: Update to SDK 2.10.1
This updates all files to be compatible with M60 and L110, primarily by
capping lower aperture to 2.37 where it is relevant. The projector
brightness is also lowered to 2.2 in order for samples to run even when
capturing with Zivid 2+ camera using white light (max projector
brightness is limited to 2.2 to keep the power consumption lower than
100W; see more info at:
https://support.zivid.com/en/latest/reference-articles/settings/
acquisition-settings/projector-brightness.html). Lastly, it makes it
possible to get intrinsics for full resolution and subsampled using
Zivid 2 cameras.