Samples: Various updates for Zivid SDK v2.16.0
Update stitching samples to use the new Zivid::UnorganizedPointCloud
APIs. This also includes adding support for Open3D for visualization.

Added a new sample to measure ambient light conditions in the scene.

Update samples that captures while projecting. We can now project
images with stronger brightness on the projector.

Reproject points sample has been simplified by using the detected
features instead of hard coding the checkerboard corners.

Samples that wait for user input with `cv::waitKey(0)` have been
updated to use a timeout value instead when built and run in CI. This
is due to a bug in the Qt backend of OpenCV on Ubuntu 24.04 which
caused the samples to not return cleanly.